Significant updates to the backend process
Add a new section on the queuing design to avoid needing to add
too much code to a stack container.  Rework and integrate the
interface contract with the pipeline into the document and be
clearer about what we're doing for the initial implementation.
Mention dramatiq-abort as the mechanism for implementing job
aborting.